+++ resources/views/bootstrap-4/components/table/row.blade.php
@@ -1,4 +1,4 @@
-@props(['url' => null, 'reordering' => false, 'customAttributes' => []])
+@props(['url' => null, 'target' => '_self', 'reordering' => false, 'customAttributes' => []])
 
 @if (!$reordering && $attributes->has('wire:sortable.item'))
     @php
@@ -10,7 +10,7 @@
     {{ $attributes->merge($customAttributes) }}
 
     @if ($url)
-        onclick="window.location='{{ $url }}';"
+        onclick="window.open('{{ $url }}', '{{ $target }}')"
         style="cursor:pointer"
     @endif
 >