+++ resources/views/bootstrap-4/includes/table.blade.php
@@ -3,6 +3,7 @@
     :customSecondaryHeader="$secondaryHeader"
     :useHeaderAsFooter="$useHeaderAsFooter"
     :customFooter="$customFooter"
+    :class="method_exists($this, 'setTableClass') ? ' ' . $this->setTableClass() : '' "
 >
     <x-slot name="head">
         @if ($reordering)
@@ -96,6 +97,7 @@
                 wire:sortable.item="{{ $row->{$primaryKey} }}"
                 :reordering="$reordering"
                 :url="method_exists($this, 'getTableRowUrl') ? $this->getTableRowUrl($row) : ''"
+                :target="method_exists($this, 'getTableRowUrlTarget') ? $this->getTableRowUrlTarget($row) : ''"
                 :class="method_exists($this, 'setTableRowClass') ? ' ' . $this->setTableRowClass($row) : ''"
                 :id="method_exists($this, 'setTableRowId') ? $this->setTableRowId($row) : ''"
                 :customAttributes="method_exists($this, 'setTableRowAttributes') ? $this->setTableRowAttributes($row) : []"