+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to `laravel-livewire-tables` will be documented in this file
 
 ## [Unreleased]
 
+## [1.19.0] - 2021-10-14
+
+### Added
+
+- Thai translation - https://github.com/rappasoft/laravel-livewire-tables/pull/491
+- Italian translation - https://github.com/rappasoft/laravel-livewire-tables/pull/493
+- Added getTableRowUrlTarget to set row click target based on the row
+- Add custom class to table - https://github.com/rappasoft/laravel-livewire-tables/pull/495
+
+### Changed
+
+- Fix removing a multiselect filter - https://github.com/rappasoft/laravel-livewire-tables/pull/494
+
 ## [1.18.0] - 2021-10-13
 
 ### Added
@@ -505,7 +518,8 @@ All notable changes to `laravel-livewire-tables` will be documented in this file
 
 - Initial release
 
-[Unreleased]: https://github.com/rappasoft/laravel-livewire-tables/compare/v1.1.0...development
+[Unreleased]: https://github.com/rappasoft/laravel-livewire-tables/compare/v1.19.0...development
+[1.19.0]: https://github.com/rappasoft/laravel-livewire-tables/compare/v1.18.0...v1.19.0
 [1.18.0]: https://github.com/rappasoft/laravel-livewire-tables/compare/v1.17.0...v1.18.0
 [1.17.0]: https://github.com/rappasoft/laravel-livewire-tables/compare/v1.16.0...v1.17.0
 [1.16.0]: https://github.com/rappasoft/laravel-livewire-tables/compare/v1.15.0...v1.16.0